slider for brush and eraser tool
