+++ invokeai/frontend/web/src/features/controlLayers/components/Settings/CanvasSettingsPopover.tsx
@@ -33,6 +33,8 @@ import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { PiCodeFill, PiEyeFill, PiGearSixFill, PiPencilFill, PiSquaresFourFill } from 'react-icons/pi';
 
+import { CanvasSettingsToolWidthSelectorDropdown } from './CanvasSettingsToolWidthSelectorDropdown';
+
 export const CanvasSettingsPopover = memo(() => {
   const { t } = useTranslation();
   return (
@@ -80,6 +82,7 @@ export const CanvasSettingsPopover = memo(() => {
               <CanvasSettingsIsolatedLayerPreviewSwitch />
               <CanvasSettingsBboxOverlaySwitch />
               <CanvasSettingsShowHUDSwitch />
+              <CanvasSettingsToolWidthSelectorDropdown />
             </Flex>
 
             <Divider />
